refactor: rename files and fix minor bugs
