+++ packages/superset-ui-chart/src/components/parseWidthOrHeight.ts
@@ -1,13 +1,13 @@
-const HUNDRED_PERCENT = { isDynamic: true, percent: 100 } as const;
+const HUNDRED_PERCENT = { isDynamic: true, multiplier: 1 } as const;
 
 export default function parseWidthOrHeight(
   input: string | number,
-): { isDynamic: true; percent: number } | { isDynamic: false; value: number } {
+): { isDynamic: true; multiplier: number } | { isDynamic: false; value: number } {
   if (input === 'auto' || input === '100%') {
     return HUNDRED_PERCENT;
   } else if (typeof input === 'string' && input.length > 0 && input[input.length - 1] === '%') {
     // eslint-disable-next-line no-magic-numbers
-    return { isDynamic: true, percent: parseFloat(input.substring(0, input.length - 1)) / 100 };
+    return { isDynamic: true, multiplier: parseFloat(input.substring(0, input.length - 1)) / 100 };
   }
   const value = typeof input === 'number' ? input : parseFloat(input);
